Refactor the logic of add entities
