+++ custom_components/terncy/__init__.py
@@ -1,4 +1,5 @@
 """The Terncy integration."""
+
 import logging
 
 from homeassistant.config_entries import ConfigEntry
