+++ custom_components/terncy/core/device.py
@@ -1,15 +1,10 @@
 import logging
 from typing import Any
 
-from homeassistant.const import (
-    CONF_DEVICE_ID,
-    CONF_DOMAIN,
-    CONF_PLATFORM,
-    CONF_TYPE,
-)
+from homeassistant.const import CONF_DEVICE_ID, CONF_DOMAIN, CONF_PLATFORM, CONF_TYPE
 
-from .entity import TerncyEntity
 from ..const import DEVICE_TRIGGER_ACTIONS_MAP, DOMAIN, HAS_EVENT_PLATFORM
+from ..hass.entity import TerncyEntity
 from ..types import AttrValue
 
 _LOGGER = logging.getLogger(__name__)
